+++ lib/utils/product_manager.dart
@@ -1,3 +1,21 @@
+/*
+********************************************************************************
+
+    _____/\\\\\\\\\_____/\\\\\\\\\\\\\\\__/\\\\\\\\\\\__/\\\\\\\\\\\\\\\_
+    ___/\\\\\\\\\\\\\__\///////\\\/////__\/////\\\///__\/\\\///////////__
+    __/\\\/////////\\\_______\/\\\___________\/\\\_____\/\\\_____________
+    _\/\\\_______\/\\\_______\/\\\___________\/\\\_____\/\\\\\\\\\\\_____
+    _\/\\\\\\\\\\\\\\\_______\/\\\___________\/\\\_____\/\\\///////______
+    _\/\\\/////////\\\_______\/\\\___________\/\\\_____\/\\\_____________
+    _\/\\\_______\/\\\_______\/\\\___________\/\\\_____\/\\\_____________
+    _\/\\\_______\/\\\_______\/\\\________/\\\\\\\\\\\_\/\\\_____________
+    _\///________\///________\///________\///////////__\///______________
+
+    Created by Muhammad Atif on 5/29/2024.
+    Portfolio https://atifnoori.web.app.
+
+ *********************************************************************************/
+
 import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_inapp_purchase/flutter_inapp_purchase.dart';
@@ -9,13 +27,16 @@ class ProductManager {
     _productIds = productIds;
   }
 
+  final iap = FlutterInappPurchase.instance;
+
   // A list to store the products items fetched from the store.
-  List<IAPItem> _productItems = [];
-  List<IAPItem> get productItems => _productItems;
+  List<ProductCommon> _productItems = [];
+
+  List<ProductCommon> get productItems => _productItems;
 
   // Method to filter items by IDs
-  List<IAPItem> getItemsByIds(List<String> ids) {
-    return _productItems.where((item) => ids.contains(item.productId)).toList();
+  List<ProductCommon> getItemsByIds(List<String> ids) {
+    return _productItems.where((item) => ids.contains(item.id)).toList();
   }
 
   // A list of products IDs for different premium features.
@@ -25,13 +46,14 @@ class ProductManager {
   Future<void> fetchProductItems() async {
     try {
       // Fetch the products items from the store using the products IDs.
-      _productItems =
-          await FlutterInappPurchase.instance.getProducts(_productIds);
+      _productItems = await iap.fetchProducts(
+        skus: _productIds,
+        type: ProductQueryType.InApp,
+      );
 
       // Sort the products items in the order of their IDs.
-      _productItems.sort((a, b) => _productIds
-          .indexOf(a.productId!)
-          .compareTo(_productIds.indexOf(b.productId!)));
+      _productItems.sort((a, b) =>
+          _productIds.indexOf(a.id).compareTo(_productIds.indexOf(b.id)));
 
       if (kDebugMode) {
         print(_productItems);